Restructure headers to be own class.
